Projects/2. qa_bot_mini-rag/pyproject.toml
[project]
name = "qa-bot-mini-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "numpy>=2.4.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "uvicorn>=0.40.0",
]
